Some steps into encapsulation
refers to #55
